Greatly simplified tests and reorganized files
Jira ticket: CAMS-396
