+++ user-interface/package.json
@@ -48,7 +48,7 @@
     "vite-plugin-static-copy": "^1.0.5",
     "vite-plugin-svgr": "^4.2.0",
     "vite-tsconfig-paths": "^4.3.2",
-    "vitest": "^1.2.2"
+    "vitest": "^1.6.0"
   },
   "eslintConfig": {
     "extends": []
