+++ user-interface/src/data-verification/consolidation/ConsolidationCasesTable.scss
