+++ user-interface/src/data-verification/consolidation/ConsolidationOrderAccordion-basics.test.tsx
@@ -1,0 +1,166 @@
+import { render, screen } from '@testing-library/react';
+import { orderType, orderStatusType } from '@/lib/utils/labels';
+import { BrowserRouter } from 'react-router-dom';
+import { ConsolidationOrder } from '@common/cams/orders';
+import {
+  ConsolidationOrderAccordion,
+  ConsolidationOrderAccordionProps,
+  fetchLeadCaseAttorneys,
+} from '@/data-verification/consolidation/ConsolidationOrderAccordion';
+import { MockData } from '@common/cams/test-utilities/mock-data';
+import { OfficeDetails } from '@common/cams/courts';
+import { formatDate } from '@/lib/utils/datetime';
+import * as FeatureFlagHook from '@/lib/hooks/UseFeatureFlags';
+import Chapter15MockApi from '@/lib/models/chapter15-mock.api.cases';
+import { FeatureFlagSet } from '@common/feature-flags';
+import { SimpleResponseData } from '@/lib/type-declarations/api';
+import { CaseAssignment } from '@common/cams/assignments';
+import { findAccordionContent, findAccordionHeading } from './testUtilities';
+
+vi.mock('../lib/components/CamsSelect', () => import('../../lib/components/CamsSelect.mock'));
+
+describe('ConsolidationOrderAccordion Basic tests', () => {
+  const order: ConsolidationOrder = MockData.getConsolidationOrder();
+  const offices: OfficeDetails[] = MockData.getOffices();
+  const regionMap = new Map();
+
+  const onOrderUpdateMockFunc = vitest.fn();
+  const onExpandMockFunc = vitest.fn();
+  let mockFeatureFlags: FeatureFlagSet;
+
+  beforeEach(async () => {
+    vi.stubEnv('CAMS_PA11Y', 'true');
+    mockFeatureFlags = {
+      'consolidations-enabled': true,
+    };
+    vitest.spyOn(FeatureFlagHook, 'default').mockReturnValue(mockFeatureFlags);
+  });
+
+  afterEach(() => {
+    vi.unstubAllEnvs();
+    vi.clearAllMocks();
+  });
+
+  function renderWithProps(props?: Partial<ConsolidationOrderAccordionProps>) {
+    const defaultProps: ConsolidationOrderAccordionProps = {
+      order,
+      officesList: offices,
+      orderType,
+      statusType: orderStatusType,
+      onOrderUpdate: onOrderUpdateMockFunc,
+      onExpand: onExpandMockFunc,
+      regionsMap: regionMap,
+    };
+
+    const renderProps = { ...defaultProps, ...props };
+    render(
+      <BrowserRouter>
+        <ConsolidationOrderAccordion {...renderProps} />
+      </BrowserRouter>,
+    );
+  }
+
+  test('should render an order heading', async () => {
+    renderWithProps();
+    const heading = findAccordionHeading(order.id!);
+    expect(heading?.textContent).toContain(order.courtName);
+    expect(heading?.textContent).toContain(formatDate(order.orderDate));
+  });
+
+  test('should display pending order content', () => {
+    const pendingOrder = MockData.getConsolidationOrder();
+    renderWithProps({ order: pendingOrder });
+    const content = findAccordionContent(pendingOrder.id!, false);
+
+    const childCaseTable = screen.getByTestId(`case-list-${pendingOrder.id}`);
+    expect(childCaseTable).toBeInTheDocument();
+
+    pendingOrder.childCases.forEach((childCase) => {
+      expect(content?.textContent).toContain(childCase.caseTitle);
+      expect(content?.textContent).toContain(formatDate(childCase.dateFiled));
+      childCase.docketEntries.forEach((de) => {
+        expect(content?.textContent).toContain(de.summaryText);
+        expect(content?.textContent).toContain(de.fullText);
+      });
+    });
+  });
+
+  test('should display approved order content', () => {
+    const leadCase = MockData.getCaseSummary();
+    const order = MockData.getConsolidationOrder({ override: { status: 'approved', leadCase } });
+    renderWithProps({ order });
+
+    const leadCaseLink = screen.queryByTestId(`lead-case-number-link`);
+    expect(leadCaseLink).toBeInTheDocument();
+
+    order.childCases.forEach((bCase, idx) => {
+      const tableRow = screen.queryByTestId(`order-${order.id}-child-cases-row-${idx}`);
+      expect(tableRow).toBeInTheDocument();
+      expect(tableRow?.textContent).toContain(bCase.caseTitle);
+    });
+  });
+
+  test('should display rejected order content', () => {
+    const order = MockData.getConsolidationOrder({
+      override: { status: 'rejected', reason: 'Test.' },
+    });
+    renderWithProps({ order });
+
+    if (order.reason) {
+      const blockQuote = document.querySelector('blockquote');
+      expect(blockQuote?.textContent).toContain(order.reason);
+    }
+
+    order.childCases.forEach((bCase, idx) => {
+      const tableRow = screen.queryByTestId(`${order.id}-case-list-row-${idx}-case-info`);
+      expect(tableRow).toBeInTheDocument();
+      expect(tableRow?.textContent).toContain(bCase.caseTitle);
+    });
+  });
+});
+
+describe('Test exported functions', () => {
+  const order: ConsolidationOrder = MockData.getConsolidationOrder();
+  let mockFeatureFlags: FeatureFlagSet;
+
+  beforeEach(async () => {
+    vi.stubEnv('CAMS_PA11Y', 'true');
+    mockFeatureFlags = {
+      'consolidations-enabled': true,
+    };
+    vitest.spyOn(FeatureFlagHook, 'default').mockReturnValue(mockFeatureFlags);
+  });
+
+  test('should return empty array when no attorneys are found', async () => {
+    vi.spyOn(Chapter15MockApi, 'get').mockImplementation((_path: string) => {
+      return Promise.resolve({
+        success: true,
+        message: '',
+        count: 1,
+        body: [],
+      } as SimpleResponseData<CaseAssignment[]>);
+    });
+
+    const attorneys = await fetchLeadCaseAttorneys(order.childCases[0].caseId);
+    expect(attorneys).toEqual([]);
+  });
+
+  test('should return string array of attorneys when found', async () => {
+    const mockAttorneys: CaseAssignment[] = MockData.buildArray(
+      () => MockData.getAttorneyAssignment(),
+      3,
+    );
+    const attorneyArray = mockAttorneys.map((assignment) => assignment.name);
+    vi.spyOn(Chapter15MockApi, 'get').mockImplementation((_path: string) => {
+      return Promise.resolve({
+        success: true,
+        message: '',
+        count: 1,
+        body: mockAttorneys,
+      } as SimpleResponseData<CaseAssignment[]>);
+    });
+
+    const attorneys = await fetchLeadCaseAttorneys(order.childCases[0].caseId);
+    expect(attorneys).toEqual(attorneyArray);
+  });
+});
