+++ user-interface/src/data-verification/transfer/TransferOrderAccordion.scss
