+++ user-interface/src/lib/utils/clone.ts
@@ -1,0 +1,3 @@
+export function deepClone(obj: object) {
+  return JSON.parse(JSON.stringify(obj));
+}
